+++ internal/models/deposit.go
@@ -7,8 +7,8 @@ type Deposit struct {
 	Name           string    `json:"name"`
 	Bank           string    `json:"bank"`
 	Type           string    `json:"type"`
-	Amount         int       `json:"amount"`
-	InitialAmount  int       `json:"initial_amount"`
+	Amount         int64     `json:"amount"`
+	InitialAmount  int64     `json:"initial_amount"`
 	InterestRate   float64   `json:"interest_rate"`
 	PromoRate      *float64  `json:"promo_rate,omitempty"`
 	PromoEndDate   string    `json:"promo_end_date,omitempty"`
